requirements.txt
# requirements.txt
google-cloud-storage==x.x.x